move run.py code to common, provide "tool" script alias, minimal tests
