+++ common/pyproject.toml
@@ -13,6 +13,9 @@ readme = "README.md"
 urls = { Homepage = "https://daveey.github.io", Repository = "https://github.com/Metta-AI/metta" }
 
 
+[project.scripts]
+tool = "metta.common.config.run_tool:main"
+
 [dependency-groups]
 dev = [
   "pyright>=1.1.401",
