+++ common/tests/config/fixtures/extra-import-root/mypackage/tools.py
@@ -1,0 +1,7 @@
+from metta.common.config.tool import Tool
+
+
+class TestTool(Tool):
+    def invoke(self, args, overrides):
+        print("TestTool invoked")
+        return 0
